pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = ["Ruslan Isupov <ruslanisupov1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = {extras = ["email"], version = "^2.10.5"}
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
cloudinary = "^1.42.1"
slowapi = "^0.1.9"
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.2"
bcrypt = "4.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
aiosqlite = "^0.20.0"
pytest-cov = "^6.0.0"
aiocache = "^0.12.3"
aioredis = "^2.0.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.1.3"

[tool.pytest.ini_options]
pythonpath = '.'
testpaths = ['tests']
filterwarnings = ["ignore::DeprecationWarning", "ignore::UserWarning"]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
timeout = 10

[tool.coverage.run]
omit = [
    "src/schemas/*",
    "src/conf/*",
    "src/database/*",
    ]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"